백엔드 소스코드 리팩토링
